+++ packages/utils/package.json
@@ -59,7 +59,7 @@
     "build:dev:watch": "yarn build:watch",
     "build:transpile:watch": "rollup -c rollup.npm.config.mjs --watch",
     "build:types:watch": "tsc -p tsconfig.types.json --watch",
-    "build:tarball": "ts-node ../../scripts/prepack.ts && npm pack ./build && ts-node ../../scripts/postpack.ts",
+    "build:tarball": "ts-node ../../scripts/pack.ts",
     "circularDepCheck": "madge --circular src/index.ts",
     "clean": "rimraf build coverage cjs esm sentry-utils-*.tgz",
     "fix": "eslint . --format stylish --fix",
